refactor(product api): extract data access logic as helper in product model
